semantic_search_lambda/requirements.txt
pgvector==0.2.4
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0
